up news

+++ src/news_window.py
@@ -116,9 +116,8 @@ class NewsGui():
                 print(self.msg.get_status())
                 return
             input_stream = session.send_finish(result)
-            os.remove(self.json_save_location)
-            #with open(self.json_save_location,"w+b") as json_f:
-             #   json_f.write(b"")
+            if os.path.exists(self.json_save_location):
+                os.remove(self.json_save_location)
             self.json_file = open(self.json_save_location,"a+b")
             input_stream.read_bytes_async(1024*500,GLib.PRIORITY_HIGH_IDLE  ,None,self.on_read_finish)
         except Exception as e:
